Fixed #1481
Moved ownership of Renderers from ControlExec to inside Visualizer class.
Added methods to delete renderers immediately or delayed, when Visualizer is
guaranteed to have active OpenGL context. If the former method is used it is
the caller's responsibility to ensure the correct OpenGL context is loaded
prior to requesting the destruction of a renderer.