Fixed #1847 - Flow: Unsteady Flow Crash
This bug was caused by an attempt to use an instance of a QuadTreeRectangle
that had previously been freed. QuadTreeRectangles are now allocated as shared
pointers to make chaching them for re-use by multiple Grid objects
easier. Populating a QuadTreeRectangle instance is an expensive operation
and it is desireable to re-use the tree as much as possible.