Issue 801 - Dataset Bug: idealized WRF data sets not supported (#2158)
* Fixed #801

* Cleanup for #801

* Fixed problem with encoding/decoding high resolution time stamps in
WRF files. WRF data sets store time in calendar form (year, month,etc) that
has to be encoded into a single value: seconds since EPOCH. For time
resolutions on the order of seconds there is not enough resolution in
a 32-bit float to encode values far from the EPOCH. This kludge works
around that problem. Ultimately, user time should be represented as a double.

* Should have been included in last commit

* Added support for WRF-SFIRE