Issue 2255 - Loading very large MPAS files results in an abort  (#2256)
* Removed duplicated Transpose code

* Made Wasp::Transpose() functions 64bit clean

* Added Initialize() method to RenderParams. Previously, there was no
initialization method and the RenderParams constructor was not no-fail.
In particular, an assert was thrown if the constructor could not determine
the extents of a variable from the DataMgr. Getting variable extents is
an operation that can fail for a number of reasons such as running out of
memory.

* Fix needed for renderers that support both 2D and 3D variables

* Fixed regression - Added Initialize() method to SliceParams, ImageParams,
and FlowParams.

* Update include/vapor/FlowParams.h

Co-Authored-By: Samuel Li <shaomeng@users.noreply.github.com>

* Update include/vapor/ImageParams.h

Co-Authored-By: Samuel Li <shaomeng@users.noreply.github.com>

* Update include/vapor/SliceParams.h

Co-Authored-By: Samuel Li <shaomeng@users.noreply.github.com>

* Fixed regression:

Even though vapor fails to open the dataset, vapor enters a limbo
state where a dataset is both open and not open. The GUI indicates
no dataset is open however if you try opening another dataset, you
will get a prompt asking whether you want to Load as new dataset
or replace existing?. I am assuming this is a bug.

* Fixing warnings generated by adding "override" qualifiers to
RenderParams::Initialize() method. The fix was to simply remove the
unused RenderParams IsOpaque() and usingVariable() methods.

* Moved GetAppRenderParams implementation to .cpp file

Co-authored-by: Samuel Li <shaomeng@users.noreply.github.com>