Fixed #2651 : Crash opening data set after a previous data set failed to open (#2673)
* Fixed #2651 : Crash opening data set after a previous data set failed to open

There are two API calls for returning the list of datasets (DataMgrs) known
to the application: ParamsMgr::GetDataMgrNames() and
DataStatus::GetDataMgrNames(). The former returns the list of data sets
contained in the application state (aka session file). The latter returns
the list of all *open* data sets. In most cases the returned values are
idential. However, if a data set has not been opened (e.g. due to failure
opening the data set) the lists will be different. There were numerous cases
in the application where the lists were assumed to be identical. I.e. if
a data set was known to the application it was assumed to be opened (have
a DataMgr associated with it). This was not the case.

Longer term it might make sense to give these API entry points more
accurately descriptive names.

* clang-format pre-push hook